created folder structure.

+++ backend/app/views/order_views.py
@@ -1,9 +1,11 @@
+from django.shortcuts import get_object_or_404
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
 from rest_framework import status
 from requests.exceptions import HTTPError
 from ..serializers import OrderSerializer
-from ..models import Order
+from ..models import Order, OrderProduct, Product
+
 
 
 # Get all orders and create new orders
@@ -15,36 +17,45 @@ def orders_list(request):
 
         serializer = OrderSerializer(data, context={'request': request}, many=True)
 
+
         return Response(serializer.data)
 
     elif request.method == 'POST':
-        
+
+        products = request.data.get('products')
+        print(products)
         serializer = OrderSerializer(data=request.data)
+
+
         if serializer.is_valid():
             serializer.save()
             return Response(status=status.HTTP_201_CREATED)
             
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+@api_view(['GET', 'PUT', 'DELETE'])
+def order_detail(request, pk):
+    
+    order = get_object_or_404(Order, pk=pk)
 
-@api_view(['DELETE'])
-def order_delete(request, pk):
-    try:
-        order = Order.objects.get(pk=pk)
-    except Order.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
+    if request.method == 'GET':
+        serializer = OrderSerializer(order)
+        return Response(serializer.data)
 
-    # if request.method == 'GET':
-    #     serializer = EncoderSerializer(encoder)
-    #     return Response(serializer.data)
+    if request.method == 'PUT':
+        serializer = OrderSerializer(order, data=request.data,context={'request': request})
 
-    # if request.method == 'PUT':
-    #     serializer = EncoderSerializer(encoder, data=request.data,context={'request': request})
-    #     if serializer.is_valid():
-    #         serializer.save()
-    #         return Response(status=status.HTTP_204_NO_CONTENT)
-    #     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    # elif request.method == 'DELETE':
-    order.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
+    elif request.method == 'DELETE':
+        order.delete()
+        
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
+def close_order(request, pk):
+    order = get_object_or_404(Order, pk=pk)
+
+
